requirements.txt
absl-py==2.4.0
accelerate==1.13.0
aenum==3.1.16
affine==2.4.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
albucore==0.0.24
albumentations==2.0.8
annotated-doc==0.0.4
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.12.1
arro3-core==0.8.0
asttokens==3.0.1
attrs==25.4.0
certifi==2026.2.25
charset-normalizer==3.4.5
ciso8601==2.3.3
click==8.3.1
cligj==0.7.2
comm==0.2.3
contourpy==1.3.3
cuda-bindings==12.9.4
cuda-pathfinder==1.4.1
cycler==0.12.1
debugpy==1.8.20
decorator==5.2.1
deltalake==1.4.2
Deprecated==1.3.1
diffusers==0.37.0
docstring_parser==0.17.0
einops==0.8.2
eotdl==2026.2.27
executing==2.2.1
filelock==3.25.0
fonttools==4.61.1
frozenlist==1.8.0
fsspec==2026.2.0
geopandas==1.1.2
gitdb==4.0.12
GitPython==3.1.46
graphviz==0.21
grpcio==1.78.0
h11==0.16.0
h5py==3.16.0
hf-xet==1.3.2
httpcore==1.0.9
httpx==0.28.1
huggingface==0.0.1
huggingface_hub==1.6.0
hydra-core==1.3.2
idna==3.11
ImageIO==2.37.2
importlib_metadata==8.7.1
importlib_resources==6.5.2
ipykernel==7.2.0
ipython==9.11.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
Jinja2==3.1.6
joblib==1.5.3
jsonargparse==4.46.0
jupyter_client==8.8.0
jupyter_core==5.9.1
kiwisolver==1.4.9
kornia==0.8.2
kornia_rs==0.1.10
lazy-loader==0.5
lightly==1.5.22
lightly-utils==0.0.2
lightning==2.6.1
lightning-utilities==0.15.3
Markdown==3.10.2
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.8
matplotlib-inline==0.2.1
mdurl==0.1.2
ml_dtypes==0.5.4
mpmath==1.3.0
mslk==0.0.0
multidict==6.7.1
nest-asyncio==1.6.0
netron==8.9.3
networkx==3.6.1
numpy==2.4.2
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.5
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvshmem-cu12==3.4.5
nvidia-nvtx-cu12==12.8.90
omegaconf==2.3.0
onnx==1.20.1
onnx-ir==0.2.0
onnxscript==0.6.2
opencv-python-headless==4.13.0.92
orjson==3.11.7
packaging==25.0
pandas==3.0.1
parso==0.8.6
pexpect==4.9.0
pillow==12.1.1
pip==26.0.1
platformdirs==4.9.4
prompt_toolkit==3.0.52
propcache==0.4.1
protobuf==6.33.5
psutil==7.2.2
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==23.0.1
pycocotools==2.0.11
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
PyJWT==2.11.0
pyogrio==0.12.1
pyparsing==3.3.2
pyproj==3.7.2
pystac==1.14.3
python-box==7.4.1
python-dateutil==2.9.0.post0
python-frontmatter==1.1.0
pytorch-lightning==2.6.1
PyYAML==6.0.3
pyzmq==27.1.0
rasterio==1.5.0
regex==2026.2.28
requests==2.32.5
rich==14.3.3
rioxarray==0.22.0
safetensors==0.7.0
scikit-image==0.26.0
scikit-learn==1.8.0
scipy==1.17.1
segmentation_models_pytorch==0.5.0
sentry-sdk==2.54.0
setuptools==80.10.2
shapely==2.1.2
shellingham==1.5.4
simsimd==6.5.16
six==1.17.0
smmap==5.0.2
stac-geoparquet==0.7.0
stack-data==0.6.3
stringzilla==4.6.0
sympy==1.14.0
tacoreader==0.5.6
tensorboard==2.20.0
tensorboard-data-server==0.7.2
termcolor==3.3.0
terratorch==1.2.2
threadpoolctl==3.6.0
tifffile==2026.3.3
timm==1.0.25
tokenizers==0.22.2
torch==2.10.0
torchao==0.16.0
torchgeo==0.9.0
torchinfo==1.8.0
torchmetrics==1.8.2
torchview==0.2.7
torchvision==0.25.0
torchvista==0.2.11
torchviz==0.0.3
tornado==6.5.4
tqdm==4.67.3
traitlets==5.14.3
transformers==5.3.0
triton==3.6.0
typer==0.24.1
typeshed_client==2.9.0
typing_extensions==4.15.0
typing-inspection==0.4.2
urllib3==2.6.3
wandb==0.25.0
wcwidth==0.6.0
Werkzeug==3.1.6
wheel==0.46.3
wrapt==2.1.2
xarray==2026.2.0
yarl==1.23.0
zipp==3.23.0
